+++ tests/mocks/mock_failing_prompt_driver.py
@@ -1,10 +1,11 @@
+from __future__ import annotations
 from collections.abc import Iterator
 from attrs import define
 
-from griptape.utils import PromptStack
-from griptape.drivers import BasePromptDriver
-from griptape.tokenizers import OpenAiTokenizer, BaseTokenizer
 from griptape.artifacts import TextArtifact
+from griptape.common import PromptStack, Message, TextMessageContent, DeltaMessage, TextDeltaMessageContent
+from griptape.drivers import BasePromptDriver
+from griptape.tokenizers import BaseTokenizer, OpenAiTokenizer
 
 
 @define
@@ -14,18 +15,25 @@ class MockFailingPromptDriver(BasePromptDriver):
     model: str = "test-model"
     tokenizer: BaseTokenizer = OpenAiTokenizer(model=OpenAiTokenizer.DEFAULT_OPENAI_GPT_3_CHAT_MODEL)
 
-    def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
+    def try_run(self, prompt_stack: PromptStack) -> Message:
         if self.current_attempt < self.max_failures:
             self.current_attempt += 1
 
-            raise Exception(f"failed attempt")
+            raise Exception("failed attempt")
         else:
-            return TextArtifact("success")
+            return Message(
+                content=[TextMessageContent(TextArtifact("success"))],
+                role=Message.ASSISTANT_ROLE,
+                usage=Message.Usage(input_tokens=100, output_tokens=100),
+            )
 
-    def try_stream(self, prompt_stack: PromptStack) -> Iterator[TextArtifact]:
+    def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]:
         if self.current_attempt < self.max_failures:
             self.current_attempt += 1
 
-            raise Exception(f"failed attempt")
+            raise Exception("failed attempt")
         else:
-            yield TextArtifact("success")
+            yield DeltaMessage(
+                content=TextDeltaMessageContent("success"),
+                usage=DeltaMessage.Usage(input_tokens=100, output_tokens=100),
+            )