+++ tests/mocks/mock_prompt_driver.py
@@ -1,11 +1,15 @@
 from __future__ import annotations
+
 from collections.abc import Iterator
 from typing import Callable
+
 from attrs import define, field
-from griptape.utils import PromptStack
+
+from griptape.artifacts import TextArtifact
+from griptape.common import PromptStack, Message, DeltaMessage, TextMessageContent, TextDeltaMessageContent
 from griptape.drivers import BasePromptDriver
 from griptape.tokenizers import BaseTokenizer
-from griptape.artifacts import TextArtifact
+
 from tests.mocks.mock_tokenizer import MockTokenizer
 
 
@@ -16,12 +20,18 @@ class MockPromptDriver(BasePromptDriver):
     mock_input: str | Callable[[], str] = field(default="mock input", kw_only=True)
     mock_output: str | Callable[[PromptStack], str] = field(default="mock output", kw_only=True)
 
-    def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
-        return TextArtifact(
-            value=self.mock_output(prompt_stack) if isinstance(self.mock_output, Callable) else self.mock_output
+    def try_run(self, prompt_stack: PromptStack) -> Message:
+        output = self.mock_output(prompt_stack) if isinstance(self.mock_output, Callable) else self.mock_output
+
+        return Message(
+            content=[TextMessageContent(TextArtifact(output))],
+            role=Message.ASSISTANT_ROLE,
+            usage=Message.Usage(input_tokens=100, output_tokens=100),
         )
 
-    def try_stream(self, prompt_stack: PromptStack) -> Iterator[TextArtifact]:
-        yield TextArtifact(
-            value=self.mock_output(prompt_stack) if isinstance(self.mock_output, Callable) else self.mock_output
-        )
+    def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]:
+        output = self.mock_output(prompt_stack) if isinstance(self.mock_output, Callable) else self.mock_output
+
+        yield DeltaMessage(content=TextDeltaMessageContent(output))
+
+        yield DeltaMessage(usage=DeltaMessage.Usage(input_tokens=100, output_tokens=100))